tests(kotlin): Add test for MonasWasmModuleProto.add

+++ wasm-module-proto/src/lib.rs
@@ -1,5 +1,5 @@
 #[no_mangle]
-pub extern "C" fn add(left: u64, right: u64) -> u64 {
+pub extern "C" fn add(left: i32, right: i32) -> i32 {
     left + right
 }
